+++ src/main/java/websockets/BotsOnDiscord.java
@@ -9,7 +9,7 @@ import org.json.JSONObject;
 
 public class BotsOnDiscord {
 
-    public static void updateServerCount(int serverCount) {
+    public static void updateServerCount(long serverCount) {
         JSONObject jsonObject = new JSONObject();
         jsonObject.put("guildCount", serverCount);
         HttpProperty[] properties = new HttpProperty[]{
