+++ src/main/java/websockets/Discordbotsgg.java
@@ -9,7 +9,7 @@ import org.json.JSONObject;
 
 public class Discordbotsgg {
 
-    public static void updateServerCount(int serverCount) {
+    public static void updateServerCount(long serverCount) {
         JSONObject jsonObject = new JSONObject();
         jsonObject.put("guildCount", serverCount);
         HttpProperty[] properties = new HttpProperty[]{
